move builder into controller module
