+++ kube-runtime/src/controller.rs
@@ -1,16 +1,22 @@
 use crate::{
-    reflector::{ErasedResource, ObjectRef, Store},
+    reflector::{
+        reflector,
+        store::{Store, Writer},
+        ErasedResource, ObjectRef,
+    },
     scheduler::{self, scheduler, ScheduleRequest},
-    utils::trystream_try_via,
+    utils::{try_flatten_applied, try_flatten_touched, trystream_try_via},
+    watcher::{self, watcher},
 };
 use derivative::Derivative;
 use futures::{
     channel, future, stream, FutureExt, SinkExt, Stream, StreamExt, TryFuture, TryFutureExt, TryStream,
     TryStreamExt,
 };
-use kube::api::Meta;
+use kube::api::{Api, ListParams, Meta};
+use serde::de::DeserializeOwned;
 use snafu::{futures::TryStreamExt as SnafuTryStreamExt, Backtrace, OptionExt, ResultExt, Snafu};
-use std::{sync::Arc, time::Duration};
+use std::{pin::Pin, sync::Arc, time::Duration};
 use tokio::time::Instant;
 
 #[derive(Snafu, Debug)]
@@ -190,4 +196,91 @@ where
     })
 }
 
-pub use crate::manager::ControllerBuilder;
+/// A builder for controller
+pub struct ControllerBuilder<K>
+where
+    K: Clone + Meta + 'static,
+{
+    // NB: Need to Unpin for stream::select_all
+    // TODO: get an arbitrary std::error::Error in here?
+    selector: Vec<Pin<Box<dyn Stream<Item = Result<ObjectRef<K>, watcher::Error>>>>>,
+    reader: Store<K>,
+}
+
+impl<K> ControllerBuilder<K>
+where
+    K: Clone + Meta + DeserializeOwned + 'static,
+{
+    /// Create a ControllerBuilder on a type `K`
+    ///
+    /// Configure `ListParams` and `Api` so you only get reconcile events
+    /// for the correct Api scope (cluster/all/namespaced), or ListParams subset
+    pub fn new(owned_api: Api<K>, lp: ListParams) -> Self {
+        let writer = Writer::<K>::default();
+        let reader = writer.as_reader();
+        let mut selector = vec![];
+        let self_watcher: Pin<Box<dyn Stream<Item = Result<ObjectRef<K>, watcher::Error>>>> = Box::pin(
+            trigger_self(try_flatten_applied(reflector(writer, watcher(owned_api, lp)))),
+        );
+        selector.push(self_watcher);
+        Self { selector, reader }
+    }
+
+    /// Retrieve a copy of the reader before starting the controller
+    pub fn store(&self) -> Store<K> {
+        self.reader.clone()
+    }
+
+    /// Indicate child objets K owns and be notified when they change
+    ///
+    /// This type `Child` must have OwnerReferences set to point back to `K`.
+    /// You can customize the parameters used by the underlying watcher if
+    /// only a subset of `Child` entries are required.
+    /// The `api` must have the correct scope (cluster/all namespaces, or namespaced)
+    pub fn owns<Child: Clone + Meta + DeserializeOwned + 'static>(
+        mut self,
+        api: Api<Child>,
+        lp: ListParams,
+    ) -> Self {
+        let child_watcher = trigger_owners(try_flatten_touched(watcher(api, lp)));
+        self.selector.push(Box::pin(child_watcher));
+        self
+    }
+
+    /// Indicate an object to watch with a custom mapper
+    ///
+    /// This mapper should return something like Option<ObjectRef<K>>
+    pub fn watches<
+        Other: Clone + Meta + DeserializeOwned + 'static,
+        I: 'static + IntoIterator<Item = ObjectRef<K>>,
+    >(
+        mut self,
+        api: Api<Other>,
+        lp: ListParams,
+        mapper: impl Fn(Other) -> I + 'static,
+    ) -> Self {
+        let other_watcher = trigger_with(try_flatten_touched(watcher(api, lp)), mapper);
+        self.selector.push(Box::pin(other_watcher));
+        self
+    }
+
+    /// Consume the ControllerBuilder and start the controller stream
+    ///
+    /// This creates a stream from all builder calls and starts a controller with
+    /// a specified `reconciler` and `error_policy` callbacks. Each of these will be called
+    /// with your configurable `Context`.
+    pub fn run<ReconcilerFut, T>(
+        self,
+        reconciler: impl FnMut(K, Context<T>) -> ReconcilerFut,
+        error_policy: impl FnMut(&ReconcilerFut::Error, Context<T>) -> ReconcilerAction,
+        context: Context<T>,
+    ) -> impl Stream<Item = Result<(ObjectRef<K>, ReconcilerAction), Error<ReconcilerFut::Error, watcher::Error>>>
+    where
+        K: Clone + Meta + 'static,
+        ReconcilerFut: TryFuture<Ok = ReconcilerAction>,
+        ReconcilerFut::Error: std::error::Error + 'static,
+    {
+        let input_stream = stream::select_all(self.selector);
+        controller(reconciler, error_policy, context, self.reader, input_stream)
+    }
+}
